Module Router fix
