+++ src/controller/cve-id.controller/cve-id.middleware.js
@@ -58,7 +58,33 @@ function parsePostParams(req, res, next) {
     next();
 }
 
+function toDate(val) {
+    let value = val.match(/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$/)
+    let result
+    if(value) {
+        result = new Date(`${value[0]}.000+00:00`)
+    } else {
+        value = val.match(/^\d{4}-\d{2}-\d{2}$/)
+        if(value) {
+            result = new Date(`${value[0]}T00:00:00.000+00:00`)
+        } else {
+            result = null
+        }
+    }
+    return result
+}
+
+function parseError(req, res, next) {
+    const errors = validationResult(req);
+    if(!errors.isempty()) {
+        res.status(400).json({ error:'BAD_INPUT', message:'Parameters were invalid', details: errors.array()})
+    }
+    next();
+}
+
 module.exports = {
     parseGetParams,
-    parsePostParams
+    parsePostParams,
+    parseError,
+    toDate
 }
