+++ src/controller/cve-id.controller/index.js
@@ -3,7 +3,7 @@ const router = express.Router()
 const mw = require('../../middleware/middleware')
 const controller = require('./cve-id.controller')
 const { check, validationResult } = require('express-validator')
-const { parseGetParams, parsePostParams } = require('./cve-id.middleware')
+const { parseGetParams, parsePostParams, parseError, toDate } = require('./cve-id.middleware')
 
 const CHOICES = ['REJECT', 'PUBLIC', 'RESERVED']
 
@@ -49,30 +49,4 @@ router.post('/cve-id-range/:year',
         parsePostParams,
         controller.CVEID_RANGE_CREATE)
 
-function toDate(val) {
-    let value = val.match(/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$/)
-    let result
-    if(value) {
-        result = new Date(`${value[0]}.000+00:00`)
-    } else {
-        value = val.match(/^\d{4}-\d{2}-\d{2}$/)
-        if(value) {
-            result = new Date(`${value[0]}T00:00:00.000+00:00`)
-        } else {
-            result = null
-        }
-    }
-    return result
-}
-
-function parseError(req, res, next) {
-    const errors = validationResult(req);
-    if(!errors.isempty()) {
-        res.status(400).json({ error:'BAD_INPUT', message:'Parameters were invalid', details: errors.array()})
-    }
-    next();
-}
-
-module.exports = {
-    router
-}
+module.exports = router
